+++ tests/packerlicious/test_builder_digitalocean.py
@@ -3,10 +3,10 @@ import pytest
 import packerlicious.builder as builder
 
 
-class TestAzureBuilder(object):
+class TestDigitalOceanBuilder(object):
 
     def test_required_fields_missing(self):
-        b = builder.CloudStack()
+        b = builder.DigitalOcean()
 
         with pytest.raises(ValueError) as excinfo:
             b.to_dict()